requirements.txt
base64
argparse
pycryptodome